requirements.txt
streamlit>=1.30.0
pandas
numpy
plotly
python-docx
reportlab
openpyxl
Pillow
pytz
requests
yfinance
fpdf2
